+++ requirements.txt
@@ -1,7 +1,7 @@
 # requirements
 # to install using pip: pip install -r requirements.txt
 
-pandas
-numpy
+pandas==1.0.3
+numpy>=1.15
 pytest
-sphinx
+sphynx
